Abstract login functionality to it's own component
