+++ LegatumClient/src/app/dashboard/dashboard.component.html
@@ -1,5 +1,5 @@
 <div id="pic-box">
-  <a routerLink="/home" ><div id="dash-signout">Sign out</div></a>
+  <a routerLink="/home" ><div id="dash-signout" (click)="logout()">Sign out</div></a>
   <h2 id="dash-greeting">Good evening, {{this.user.username}}</h2>
 </div>
 <div id="mid-nav">
